requirements.txt
selenium~=4.23.1
Django~=5.0.7
djangorestframework~=3.15.2